requirements.txt
# Core dependencies
requests>=2.28.0
pandas>=1.5.0
openpyxl>=3.0.10
python-dotenv>=0.19.0

# Data processing
numpy>=1.21.0
jsonschema>=4.0.0

# Excel handling
xlsxwriter>=3.0.0

# API and web
urllib3>=1.26.0
certifi>=2022.0.0

# MCP (Model Context Protocol)
mcp>=0.5.0

# Optional: Advanced features
Pillow>=9.0.0  # For image processing
matplotlib>=3.5.0  # For analytics visualization

# Development dependencies (install with: pip install -r requirements-dev.txt)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0
# mypy>=0.950

